:recycle: Refactor: Profile 컴포넌트 모듈화 작업
